requirements.txt
djangorestframework==3.14.0
psycopg2==2.9.6
Pillow==9.5.0
django-import-export==3.2.0
django-filter==23.2
django-ckeditor==6.6.1
django-cors-headers==4.2.0
secure-smtplib==0.1.1
requests==2.32.3
django-environ==0.10.0
